tests: consolidate the search suite and mirror its source subpackages
`tests/compiler/pipeline/search/` held 30 flat test modules against a source
package that has had `data/`, `policy/`, and `prior/` subpackages for a while.
Two changes, no behavior difference.

Consolidation — merge the file groups that re-declared each other's fixtures:

- The golden tier's three files become one `policy/test_golden_evidence.py`.
  `test_golden_floor.py` duplicated CARD / CAP / _SIG / _STD_TILE / _W1X1_TILE /
  _ROW_W1X1 / _ROW_GOLD / _golden() / _FakeFP verbatim from the evidence file;
  `test_golden_evidence_real_resolve.py` is the same tier through a live resolve.
  They are now three sections against one set of q_proj fixtures.
- The three `OfflinePrior` files become `prior/test_offline_prior.py` — the
  weights artifact, the quality proxy, and the tensor-core gates, one subject.

`test_golden_deploy_perf.py` moved beside `test_diagnostics.py` but was NOT
merged into it: its `_single_gpu_goldens` fixture is autouse, so folding it in
would have silently applied single-card golden scoping to every diagnostics test.

`test_db_evidence.py`, `test_deploy_pick_determinism.py` and
`test_golden_spelling_canonical.py` stay separate — they share no scaffolding
with the golden tier and cover different subjects (the DB tier, a cross-tier
order-invariance property, and the stored-corpus tripwire).

Layout — tests move into subpackages matching the source:

  data/    <- data/dataset.py, data/shape.py, data/freeze.py
  policy/  <- policy/greedy.py, policy/mcts.py
  prior/   <- prior/online.py, prior/offline.py, prior/diagnostics.py, prior/fit/

Tests of the package's own top-level modules (db, features, keys, slice, …) and
the cross-cutting ones (`test_boot_resolution_cache.py`, `test_structural_push.py`)
stay flat. `conftest.py` stays at the package root and reaches every subdirectory.

Docs: `tests/ARCHITECTURE.md` drops the exhaustive per-file tree and the two
per-file unit/CLI tables — both are the rot-prone listings the repo's own
documentation conventions forbid — for a structural map plus the mirroring rule
and a "one file per subject, not per bug" convention.
