+++ tests/compiler/pipeline/search/data/test_freeze.py
