+++ tests/compiler/pipeline/search/policy/test_deploy_pick_determinism.py
