+++ tests/compiler/pipeline/search/prior/test_diagnostics.py
