+++ tests/compiler/pipeline/search/prior/test_golden_deploy_perf.py
